Took Jones Advice. Converted to Bootstrap Theme.
It looks good. For the main result page, let me know if you think otherwise.

We need to work on the following .html and css of the following pages:
1. login.html
2. upload_application.html
3. sign_up.html
4. college.html

We will be using php for all of these but for now if we create the static HTML pages it will allow for us to know what to do once we start writing our PhP scripts.