login, register, forgot password pages are done
HTML that needs work:
Upload_application.html
search.html (This will be the initial page once we are in production)

PHP work HTML generating:
But top_nav_bar into php file so we can call it on each page.
put main_results in index into a php page so they can become dynamic 
refine bar in index into php so it can be dynamic. (This is a maybe and is not as imporant)

PHP scripts for getting data:
PHP script for MYSQL server, PhP function to SELECT, UPDATE, INSERT, connect, disconnect.
Script to determine what information to display. (This will be performed after the query)